API Changes and Additions
Lots of things changed. For one, a PC that displays all unique variable names was drafted. Also, the process by which files are chosen for merging. The new process allows the user to backtrack without losing previous selections.  Some code cleanup was performed. Some method and class summaries were drafted. Autocomplete was added to the filter selection PC, and will be added to all drop down lists for user convenience.